+++ lizard_area/geoobject.py
@@ -9,13 +9,14 @@ from django.template.defaultfilters import slugify
 from django.contrib.gis.geos import GEOSGeometry
 from osgeo import ogr
 
-from lizard_area.models import Communique
+from lizard_area.models import Area
+from lizard_area.models import AreaCode
 from lizard_area.models import GeoObjectGroup
 
 logger = logging.getLogger(__name__)
 
 
-def import_shapefile_communique(shapefile_filename, user):
+def import_shapefile_area(shapefile_filename, user, data_administrator):
     """
     Load shapefile with communique and put it in Communique(GeoObject)
     and GeoObjectGroup
@@ -46,6 +47,8 @@ def import_shapefile_communique(shapefile_filename, user):
         created_by=user)
     geo_object_group.save()
 
+    done = {}  # Check for duplicates
+
     # Add all geo objects to group.
     for feature in layer:
         geom = feature.GetGeometryRef()
@@ -54,26 +57,71 @@ def import_shapefile_communique(shapefile_filename, user):
         # geom.FlattenTo2D()
         # import pdb;pdb.set_trace()
 
-        kwargs = {
-            'ident': feature.GetField(feature.GetFieldIndex('OWAIDENT')),
-            'geometry': GEOSGeometry(geom.ExportToWkt(), srid=4326),
-            'geo_object_group': geo_object_group,
+        # # KRW Waterlichamen merge
+        # kwargs = {
+        #     'ident': feature.GetField(feature.GetFieldIndex('OWAIDENT')),
+        #     'geometry': GEOSGeometry(geom.ExportToWkt(), srid=4326),
+        #     'geo_object_group': geo_object_group,
 
-            # Communique
-            'name': '%s ' % feature.GetField(feature.GetFieldIndex('OWANAAM'))
-        }
-        # Afvoergebieden
+        #     # Communique
+        #     'name': '%s' % feature.GetField(feature.GetFieldIndex('OWANAAM')),
+
+        #     # Area
+        #     'data_administrator': data_administrator,
+        #     'area_class': Area.AREA_CLASS_KRW_WATERLICHAAM,
+        # }
+
+        # # Afvoergebieden, 2 varianten
+        # # area_code, _ = AreaCode.objects.get_or_create(
+        # #     name=feature.GetField(feature.GetFieldIndex('GAFCODE')))
         # kwargs = {
         #     'ident': feature.GetField(feature.GetFieldIndex('GAFIDENT')),
         #     'geometry': GEOSGeometry(geom.ExportToWkt(), srid=4326),
         #     'geo_object_group': geo_object_group,
 
         #     # Communique
-        #     'name': '%s ' % feature.GetField(feature.GetFieldIndex('GAFNAAM')),
-        #     'code': feature.GetField(feature.GetFieldIndex('GAFCODE')),
+        #     'name': unicode(
+        #         feature.GetField(feature.GetFieldIndex('GAFNAAM')),
+        #         errors='ignore'),
+        #     #'code': area_code,
+        #     # Area
+        #     'data_administrator': data_administrator,
+        #     'area_class': Area.AREA_CLASS_AAN_AFVOERGEBIED,
         # }
-        geo_object = Communique(**kwargs)
-        geo_object.save()
+
+        # Deel afvoergebieden: watervlakken_mbp_sap
+        # area_code, _ = AreaCode.objects.get_or_create(
+        #     name=feature.GetField(feature.GetFieldIndex('GAFCODE')))
+        parent = Area.objects.get(name=unicode(
+                feature.GetField(feature.GetFieldIndex('GAFNAAM')),
+                errors='ignore'))
+        # Only field that is unique.
+        name = 'Deelgebied %s' % feature.GetField(
+            feature.GetFieldIndex('FID_GBKNwa'))
+        ident = 'LSDR-%s' % feature.GetField(
+            feature.GetFieldIndex('FID_GBKNwa')),  # Else it is not unique
+        kwargs = {
+            'ident': ident,
+            'geometry': GEOSGeometry(geom.ExportToWkt(), srid=4326),
+            'geo_object_group': geo_object_group,
+
+            # Communique
+            'name': name,
+            #'code': area_code,
+            # Area
+            'data_administrator': data_administrator,
+            'area_class': Area.AREA_CLASS_DEEL_AAN_AFVOERGEBIED,
+            'parent': parent,
+        }
+
+        if ident not in done:
+            geo_object = Area(**kwargs)
+            geo_object.save()
+            done[ident] = None
+        else:
+            logger.warning(
+                'Ignored: Ident %s is already imported in this session' %
+                ident)
         number_of_features += 1
 
     logger.info("Added %s with %s geo objects.",