+++ lizard_area/models.py
@@ -2,60 +2,13 @@
 import logging
 from treebeard.al_tree import AL_Node
 
-from django.contrib.auth.models import User
 from django.contrib.gis.db import models
 from django.core.urlresolvers import reverse
 
+from lizard_geo.models import GeoObject
+from lizard_geo.models import GeoObjectGroup
+
 logger = logging.getLogger(__name__)
-
-
-class GeoObjectGroup(models.Model):
-    """
-    Geo objects are grouped.
-
-    These are starting points to navigate through all geo objects.
-
-    Examples of groups:
-    - Alle aan-/afvoergebieden van een waterschap
-    - Alle deel aan-/afvoergebieden van 1 deelgebied
-    - Alle krw waterlichamen
-
-    TODO: Automatically fill in slug
-    """
-    name = models.CharField(max_length=128)
-    slug = models.SlugField(unique=True)
-    # legend = models.ForeignKey(LegendClass, null=True, blank=True)
-    # "source"
-
-    created_by = models.ForeignKey(User)
-    last_modified = models.DateTimeField(auto_now=True)
-
-    def __unicode__(self):
-        return self.name
-
-    def get_absolute_url(self):
-        return reverse('lizard-area:api:geo_object_group',
-                       kwargs={'pk': self.pk})
-
-
-class GeoObject(AL_Node):
-    """
-    Geo objects storage.
-
-    Ident MUST be unique. When importing shapefiles references are
-    done using ident.
-
-    Parents are used for deel aan-/afvoergebieden.
-    """
-    ident = models.CharField(max_length=16, unique=True)
-    geometry = models.GeometryField(srid=4326)
-    geo_object_group = models.ForeignKey(GeoObjectGroup)
-    objects = models.GeoManager()
-
-    # verwijzing naar "bron locatie" string
-
-    def __unicode__(self):
-        return '%s (%s)' % (self.ident, self.geo_object_group)
 
 ############################
 
@@ -119,23 +72,66 @@ class Category(AL_Node):
         return '%s (%s)' % (self.name, self.slug)
 
     def get_absolute_url(self):
-        return reverse('lizard-area:api:category', kwargs={'slug': self.slug})
+        return reverse('lizard_area_api_category', kwargs={'slug': self.slug})
 
 
 #############################
 
-class AreaAdministrator(models.Model):
+class NameAbstract(models.Model):
     """
-    Administrators of areas.
-
-    The administrator can be an organization.
+    Abstract model with only a name as property.
     """
     name = models.CharField(max_length=128, unique=True)
-    geo_object_groups = models.ManyToManyField(
-        GeoObjectGroup, null=True, blank=True)
+
+    class Meta:
+        abstract = True
 
     def __unicode__(self):
         return self.name
+
+
+class DataAdministrator(NameAbstract):
+    """
+    Administrators of all kinds of data.
+
+    The administrator can be an organization.
+    """
+    pass
+
+
+class AreaCode(NameAbstract):
+    """Area code, used by communique."""
+    pass
+
+
+class Status(NameAbstract):
+    """Area status, used by communique."""
+    pass
+
+
+class AreaType(NameAbstract):
+    """Area type, used by communique."""
+    pass
+
+
+class Province(NameAbstract):
+    """Province, used by communique."""
+    pass
+
+
+class Municipality(NameAbstract):
+    """Municipality, used by communique."""
+    pass
+
+
+class Basin(NameAbstract):
+    """Basin ("stroomgebied"), used by communique."""
+    pass
+
+
+class WaterManagementArea(NameAbstract):
+    """Water management area ("waterbeheergebied"), used by communique."""
+    pass
 
 
 class Communique(GeoObject):
@@ -150,29 +146,65 @@ class Communique(GeoObject):
     name = models.CharField(max_length=128)
 
     # i.e. "NL13_11"
-    code = models.CharField(max_length=128, null=True, blank=True)
+    code = models.ForeignKey(AreaCode, null=True, blank=True)
 
     # i.e. "Kunstmatig"
-    status = models.CharField(max_length=128, null=True, blank=True)
+    status = models.ForeignKey(Status, null=True, blank=True)
 
     # i.e. "M27 - Matig grote ondiepe laagveenplassen"
-    area_type = models.CharField(max_length=128, null=True, blank=True)
+    area_type = models.ForeignKey(AreaType, null=True, blank=True)
 
     # i.e. "Zuid-Holland"
-    province = models.CharField(max_length=128, null=True, blank=True)
+    province = models.ForeignKey(Province, null=True, blank=True)
 
     # i.e. "Reeuwijk"
-    municipality = models.CharField(max_length=128, null=True, blank=True)
+    municipality = models.ForeignKey(Municipality, null=True, blank=True)
 
     # Stroomgebied. i.e. "Rijn-West"
-    basin = models.CharField(max_length=128, null=True, blank=True)
+    basin = models.ForeignKey(Basin, null=True, blank=True)
 
     # Waterbeheergebied. i.e. "Hoogheemraadschap van Rijnland"
-    watermanagementarea = models.CharField(
-        max_length=128, null=True, blank=True)
+    watermanagementarea = models.ForeignKey(
+        WaterManagementArea, null=True, blank=True)
 
     def __unicode__(self):
         return '%s - %s' % (self.ident, self.name)
 
     def get_absolute_url(self):
-        return reverse('lizard-area:api:communique', kwargs={'pk': self.pk})
+        return reverse('lizard_area_api_communique', kwargs={'pk': self.pk})
+
+
+class Area(Communique, AL_Node):
+    """
+    KRW waterlichamen en (deel) aan-/afvoergebieden.
+    """
+
+    AREA_CLASS_KRW_WATERLICHAAM = 1
+    AREA_CLASS_AAN_AFVOERGEBIED = 2
+    AREA_CLASS_DEEL_AAN_AFVOERGEBIED = 3
+
+    AREA_CLASS_CHOICES = (
+        (AREA_CLASS_KRW_WATERLICHAAM, 'krw waterlichaam'),
+        (AREA_CLASS_AAN_AFVOERGEBIED, 'aan-/afvoer gebied'),
+        (AREA_CLASS_DEEL_AAN_AFVOERGEBIED, 'deel aan-/afvoergebied'), )
+
+    AREA_CLASS_DICT = dict(AREA_CLASS_CHOICES)
+
+    parent = models.ForeignKey('Area', null=True, blank=True)
+    data_administrator = models.ForeignKey(DataAdministrator)
+    area_class = models.IntegerField(
+        choices=AREA_CLASS_CHOICES, default=AREA_CLASS_KRW_WATERLICHAAM)
+
+    # For treebeard.
+    node_order_by = ['name']
+
+    class Meta:
+        ordering = ('name', )
+
+    def __unicode__(self):
+        return '%s (%s - %s)' % (
+            self.name, self.AREA_CLASS_DICT[self.area_class],
+            self.data_administrator)
+
+    def get_absolute_url(self):
+        return reverse('lizard_area_api_area', kwargs={'pk': self.pk})