+++ lizard_area/testsettings.py
@@ -47,7 +47,9 @@ DATABASES = {
 SITE_ID = 1
 INSTALLED_APPS = [
     'lizard_area',
+    'lizard_geo',
     'lizard_ui',
+    'lizard_map',
     'treebeard',
     'djangorestframework',
     'staticfiles',