+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog of lizard-area
 0.1 (unreleased)
 ----------------
 
+- Created geoobjects and moved it to lizard-geo.
+
+- Created initial tests.
+
+- Created initial migration.
+
 - Created initial adapter.
 
 - Created initial api.